+++ papis/commands/browse.py
@@ -71,7 +71,7 @@ from typing import Optional
 logger = logging.getLogger('browse')
 
 
-def run(document: papis.document.Document) -> Optional[str]:
+def run(document: papis.document.Document, browse: bool=True) -> Optional[str]:
     """Browse document's url whenever possible and returns the url
 
     :document: Document object
@@ -106,8 +106,11 @@ def run(document: papis.document.Document) -> Optional[str]:
                    + '/?'
                    + urlencode(params))
 
-    logger.info("Opening url %s:" % url)
-    papis.utils.general_open(url, "browser", wait=False)
+    if browse:
+        logger.info("Opening url %s:" % url)
+        papis.utils.general_open(url, "browser", wait=False)
+    else:
+        print(url)
     return url
 
 
@@ -118,11 +121,14 @@ def run(document: papis.document.Document) -> Optional[str]:
 @click.option('-k', '--key', default='',
               help='Use the value of the document\'s key to open in'
                    ' the browser, e.g. doi, url, doc_url ...')
+@click.option('-n', '--print', "_print", default=False, is_flag=True,
+              help='Just print out the url, do not open it with browser')
 @papis.cli.all_option()
 @papis.cli.doc_folder_option()
 def cli(query: str,
         key: str,
         _all: bool,
+        _print: bool,
         doc_folder: str,
         sort_field: Optional[str],
         sort_reverse: bool) -> None:
@@ -153,4 +159,4 @@ def cli(query: str,
     logger.info("Using key = %s", papis.config.get("browse-key"))
 
     for document in documents:
-        run(document)
+        run(document, browse=not _print)